[feat] 아파트 기본 정보 컴포넌트 분리 #36
