+++ src/components/Main/ApartInfo/ApartInfo.tsx
@@ -1,3 +1,5 @@
+// 아파트 기본정보 컴포넌트
+
 import styles from "./ApartInfo.module.scss";
 import detailButtonIconD from "@/assets/Main/ApartInfo/detailButtonIconD.svg";
 import ApartSearch from "@/components/Common/ApartSearch/ApartSearch";
@@ -5,7 +7,7 @@ import ApartHeadInfo from "./ApartHeadInfo";
 import ApartBuildInfo from "./ApartMainInfo/ApartBuildInfo/ApartBuildInfo";
 import ApartPrice from "./ApartMainInfo/ApartPrice";
 import ApartMaintanceCharge from "./ApartMainInfo/ApartMaintanceCharge";
-import ApartGeneralInfo from "./ApartMainInfo/ApartGeneralInfo";
+import ApartGeneralInfo from "./ApartMainInfo/ApartGeneralInfo/ApartGeneralInfo";
 import { useApartInfoStore } from "@/store/useApartInfoStore";
 import { useMainInfoStore } from "@/store/useMainInfoStore";
 import getApartData from "@/utils/api/getApartData";
