--- src/components/Main/ApartInfo/ApartMainInfo/ApartGeneralInfo.module.scss
@@ -1,62 +1,0 @@
-@use "@/styles/_colors.scss";
-
-.apartGeneralInfo {
-  display: grid;
-  grid-template-columns: repeat(2, minmax(0, 1fr));
-  grid-template-rows: auto auto;
-  gap: 0.8rem;
-  li {
-    &:first-child {
-      grid-column: 1 / -1;
-    }
-    .title {
-      padding: 1rem 2.6rem 1rem;
-      font-size: 1.8rem;
-      font-weight: 600;
-      display: flex;
-      align-items: center;
-      gap: 0.8rem;
-      color: #ffffff;
-      background-color: colors.$mainTitle;
-      img {
-        width: 2rem;
-        height: 2rem;
-      }
-    }
-    .amenities,
-    .content {
-      display: inline-block;
-      width: 100%;
-      padding: 0.8rem 2.4rem 0.8rem;
-      border: 0.1rem solid #9e9e9e;
-      border-top: none;
-      color: colors.$mainNavy;
-      font-size: 1.6rem;
-      text-overflow: ellipsis;
-      overflow: hidden;
-      white-space: nowrap;
-      li {
-        display: flex;
-        align-items: center;
-        gap: 0.4rem;
-        span {
-          color: colors.$mainNavy;
-          font-weight: 500;
-        }
-      }
-    }
-    .amenities {
-      display: flex;
-      flex-wrap: wrap;
-      gap: 1rem;
-    }
-  }
-}
-
-.tooltip {
-  z-index: 10;
-  background: white;
-  color: #142337;
-  border: 0.2rem solid #e03a3e;
-  font-size: 1.4rem;
-}
