+++ src/components/Main/ApartInfo/ApartMainInfo/ApartGeneralInfo/ApartGeneralInfo.module.scss
@@ -1,0 +1,6 @@
+.apartGeneralInfo {
+  display: grid;
+  grid-template-columns: repeat(2, minmax(0, 1fr));
+  grid-template-rows: auto auto;
+  gap: 0.8rem;
+}
